requirements.txt
flask
faiss-cpu
sentence-transformers
transformers
torch
requests
bs4
PyMuPDF
